+++ modules/canvaskit/canvaskit_bindings.cpp
@@ -175,6 +175,11 @@ void ApplyAddArc(SkPath& orig, const SkRect& oval, SkScalar startAngle, SkScalar
     orig.addArc(oval, startAngle, sweepAngle);
 }
 
+void ApplyAddOval(SkPath& orig, const SkRect& oval, bool ccw, unsigned start) {
+    orig.addOval(oval, ccw ? SkPath::Direction::kCCW_Direction :
+                             SkPath::Direction::kCW_Direction, start);
+}
+
 void ApplyAddPath(SkPath& orig, const SkPath& newPath,
                    SkScalar scaleX, SkScalar skewX,  SkScalar transX,
                    SkScalar skewY,  SkScalar scaleY, SkScalar transY,
@@ -211,6 +216,13 @@ void ApplyArcTo(SkPath& p, SkScalar x1, SkScalar y1, SkScalar x2, SkScalar y2,
 
 void ApplyArcToAngle(SkPath& p, SkRect& oval, SkScalar startAngle, SkScalar sweepAngle, bool forceMoveTo) {
     p.arcTo(oval, startAngle, sweepAngle, forceMoveTo);
+}
+
+void ApplyAddArcToArcSize(SkPath& orig, SkScalar rx, SkScalar ry, SkScalar xAxisRotate,
+                          bool useSmallArc, bool ccw, SkScalar x, SkScalar y) {
+    auto arcSize = useSmallArc ? SkPath::ArcSize::kSmall_ArcSize : SkPath::ArcSize::kLarge_ArcSize;
+    auto sweep = ccw ? SkPath::Direction::kCCW_Direction : SkPath::Direction::kCW_Direction;
+    orig.arcTo(rx, ry, xAxisRotate, arcSize, sweep, x, y);
 }
 
 void ApplyClose(SkPath& p) {
@@ -754,6 +766,7 @@ EMSCRIPTEN_BINDINGS(Skia) {
             }
             self.drawAtlas(atlas, dstXforms, srcRects, colors, count, mode, nullptr, paint);
         }), allow_raw_pointers())
+        .function("drawCircle", select_overload<void (SkScalar, SkScalar, SkScalar, const SkPaint& paint)>(&SkCanvas::drawCircle))
         .function("drawImage", select_overload<void (const sk_sp<SkImage>&, SkScalar, SkScalar, const SkPaint*)>(&SkCanvas::drawImage), allow_raw_pointers())
         .function("drawImageRect", optional_override([](SkCanvas& self, const sk_sp<SkImage>& image,
                                                         SkRect src, SkRect dst,
@@ -791,6 +804,7 @@ EMSCRIPTEN_BINDINGS(Skia) {
         .function("drawTextBlob", select_overload<void (const sk_sp<SkTextBlob>&, SkScalar, SkScalar, const SkPaint&)>(&SkCanvas::drawTextBlob))
         .function("drawVertices", select_overload<void (const sk_sp<SkVertices>&, SkBlendMode, const SkPaint&)>(&SkCanvas::drawVertices))
         .function("flush", &SkCanvas::flush)
+        .function("getSaveCount", &SkCanvas::getSaveCount)
         .function("getTotalMatrix", optional_override([](const SkCanvas& self)->SimpleMatrix {
             SkMatrix m = self.getTotalMatrix();
             return toSimpleSkMatrix(m);
@@ -968,6 +982,7 @@ EMSCRIPTEN_BINDINGS(Skia) {
         .constructor<const SkPath&>()
         .function("_addArc", &ApplyAddArc)
         // interface.js has 3 overloads of addPath
+        .function("_addOval", &ApplyAddOval)
         .function("_addPath", &ApplyAddPath)
         // interface.js has 4 overloads of addRect
         .function("_addRect", &ApplyAddRect)
@@ -975,6 +990,7 @@ EMSCRIPTEN_BINDINGS(Skia) {
         .function("_addRoundRect", &ApplyAddRoundRect)
         .function("_arcTo", &ApplyArcTo)
         .function("_arcTo", &ApplyArcToAngle)
+        .function("_arcTo", &ApplyAddArcToArcSize)
         .function("_close", &ApplyClose)
         .function("_conicTo", &ApplyConicTo)
         .function("countPoints", &SkPath::countPoints)